Finished migrating tests for score.ExtractResults
