+++ kit/score/results_test.go
@@ -1,4 +1,4 @@
-package ci
+package score_test
 
 import (
 	"encoding/json"
@@ -8,8 +8,6 @@ import (
 	"time"
 
 	"github.com/autograde/quickfeed/kit/score"
-	"github.com/autograde/quickfeed/log"
-	"go.uber.org/zap"
 )
 
 func TestExtractResult(t *testing.T) {
@@ -20,7 +18,7 @@ func TestExtractResult(t *testing.T) {
 Here are some more logs for the student.
 `
 
-	res, err := ExtractResult(zap.NewNop().Sugar(), out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
+	res, err := score.ExtractResults(out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -38,7 +36,7 @@ func TestExtractResultWithWhitespace(t *testing.T) {
 Here are some more logs for the student.
 `
 
-	res, err := ExtractResult(zap.NewNop().Sugar(), out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
+	res, err := score.ExtractResults(out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -61,7 +59,7 @@ Here are some more logs for the student.
 Here are some more logs for the student.
 `
 
-	res, err := ExtractResult(zap.NewNop().Sugar(), out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
+	res, err := score.ExtractResults(out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -84,13 +82,16 @@ func TestExtractResultWithPanicedAndMaliciousScoreLines(t *testing.T) {
 	{"Secret":"59fd5fe1c4f741604c1beeab875b9c789d2a7c73","TestName":"MaliciousTest","Score":100,"MaxScore":100,"Weight":1}
 `
 
-	res, err := ExtractResult(zap.NewNop().Sugar(), out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
+	res, err := score.ExtractResults(out, "59fd5fe1c4f741604c1beeab875b9c789d2a7c73", 10)
 	if err != nil {
 		t.Fatal(err)
 	}
-	const expected = 6
-	if len(res.ScoreMap) != expected {
-		t.Fatalf("ExtractResult() expected %d Score entries, got %d: %+v", expected, len(res.ScoreMap), res.ScoreMap)
+	const expectedTests = 6
+	if len(res.ScoreMap) != expectedTests {
+		t.Fatalf("ExtractResult() expected %d Score entries, got %d: %+v", expectedTests, len(res.ScoreMap), res.ScoreMap)
+	}
+	if len(res.TestNames) != expectedTests {
+		t.Fatalf("Extract() expected %d Test entries, got %d: %+v", expectedTests, len(res.TestNames), scores)
 	}
 
 	testOrder := []string{
@@ -114,22 +115,34 @@ var scoreObjects = `
 [{"Secret":"hidden","TestName":"TestLintAG","Score":3,"MaxScore":3,"Weight":5},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/Two_jobs","Score":2,"MaxScore":2,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/Three_jobs","Score":3,"MaxScore":3,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/Five_jobs","Score":5,"MaxScore":5,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/Six_jobs","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/FIFO/Six_jobs_unordered","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/Two_jobs","Score":10,"MaxScore":10,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/Three_jobs","Score":15,"MaxScore":15,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/Five_jobs","Score":25,"MaxScore":25,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/Six_jobs","Score":28,"MaxScore":28,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(2)/Six_jobs_unordered","Score":28,"MaxScore":28,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/Two_jobs","Score":4,"MaxScore":4,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/Three_jobs","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/Five_jobs","Score":10,"MaxScore":10,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/Six_jobs","Score":12,"MaxScore":12,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(5)/Six_jobs_unordered","Score":12,"MaxScore":12,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/Two_jobs","Score":2,"MaxScore":2,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/Three_jobs","Score":3,"MaxScore":3,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/Five_jobs","Score":5,"MaxScore":5,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/Six_jobs","Score":8,"MaxScore":8,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/RR(10)/Six_jobs_unordered","Score":8,"MaxScore":8,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Two_jobs","Score":2,"MaxScore":2,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Three_jobs","Score":3,"MaxScore":3,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Five_jobs","Score":5,"MaxScore":5,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Six_jobs","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Six_jobs_unordered","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SJF/Six_jobs_different_unordered","Score":6,"MaxScore":6,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SS(5)/No_jobs","Score":0,"MaxScore":0,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SS(5)/ABC_jobs","Score":12,"MaxScore":12,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SS(5)/ABC_jobs_long","Score":60,"MaxScore":60,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SS(5)/Varying_length_ABC_jobs","Score":32,"MaxScore":32,"Weight":2},{"Secret":"hidden","TestName":"TestSchedulersAG/SS(5)/ABCDE_jobs","Score":84,"MaxScore":84,"Weight":2}]
 `
 
-func TestTotalScore(t *testing.T) {
-	result := &score.Result{}
-	result.Scores = make([]*score.Score, 0)
+func TestScoresSum(t *testing.T) {
+	scores := make([]*score.Score, 0)
 	dec := json.NewDecoder(strings.NewReader(scoreObjects))
 	for {
-		if err := dec.Decode(&result.Scores); err == io.EOF {
+		if err := dec.Decode(&scores); err == io.EOF {
 			break
 		} else if err != nil {
 			t.Fatal(err)
 		}
 	}
-	got := result.TotalScore()
+	s := score.NewScores()
+	const hiddenSecret = "hidden"
+	for _, sc := range scores {
+		// The scoreObjects was extracted when we allowed Weight=0
+		// We now return an error for when Weight=0.
+		// Hence, we only add scores with non-zero weights.
+		if err := sc.IsValid(hiddenSecret); err == nil {
+			s.AddScore(sc)
+		}
+	}
+	err := s.Validate(hiddenSecret)
+	if err != nil {
+		t.Errorf("Validate() = %v, expected <nil>", err)
+	}
+	got := s.Sum()
 	const want = 100
 	if got != want {
-		t.Errorf("TotalScore() = '%d', want '%d'", got, want)
-		t.Logf("scores: %+v", log.IndentJson(result.Scores))
+		t.Errorf("Sum() = '%d', want '%d'", got, want)
 	}
 }
 
@@ -145,10 +158,9 @@ func TestExecTime(t *testing.T) {
 		{"4", 2_800_000_000, 2800},
 		{"5", 3_888_900_000, 3888},
 	}
-	logger := zap.NewNop().Sugar()
 	for _, tt := range tests {
 		t.Run("ExecTime#"+tt.id, func(t *testing.T) {
-			res, err := ExtractResult(logger, "", "", tt.in)
+			res, err := score.ExtractResults("", "", tt.in)
 			if err != nil {
 				t.Fatal(err)
 			}
